+++ mlx_engine/generate.py
@@ -27,6 +27,7 @@ from mlx_engine.utils.speculative_decoding import (
 from outlines.processors.structured import JSONLogitsProcessor
 from mlx_engine.utils.outlines_transformer_tokenizer import OutlinesTransformerTokenizer
 from mlx_engine.cache_wrapper import StopPromptProcessing, PROMPT_PROCESSING_CHUNK_SIZE
+from mlx_engine.utils.progress_decorators import ratchet, throw_to_stop, token_count
 
 MAX_TOP_LOGPROBS = 10
 
@@ -122,97 +123,6 @@ def unload_draft_model(model_kit: ModelKit | VisionModelKit) -> None:
     model_kit.unload_draft_model()
 
 
-def ratchet_callback_wrapper(callback: Optional[Callable[[float], bool]]):
-    """
-    Wraps a progress callback to ensure progress values are monotonically increasing.
-
-    This wrapper prevents progress from appearing to move backwards by using a ratchet
-    mechanism. If a lower percentage is reported than previously seen, the callback
-    returns True (continue) without calling the original callback.
-
-    Args:
-        callback: A callback that accepts progress (0.0–100.0) and returns
-                 True to continue or False to stop. May be None.
-
-    Returns:
-        A wrapped callback that ensures monotonic progress reporting.
-        If callback is None, returns None.
-    """
-    if callback is None:
-        return None
-
-    ratchet = -1.0
-
-    def inner_callback(percentage: float) -> bool:
-        nonlocal ratchet
-        if percentage <= ratchet:
-            return True
-        ratchet = percentage
-        return callback(percentage)
-
-    return inner_callback
-
-
-def throw_to_stop_callback_wrapper(callback: Optional[Callable[[float], bool]]):
-    """
-    Wraps a progress callback to raise an exception when stopping is requested.
-
-    Instead of returning False to indicate stopping, this wrapper raises a
-    StopPromptProcessing exception when the original callback returns False.
-    This allows for immediate termination of the generation process.
-
-    Args:
-        callback: A callback that accepts progress (0.0–100.0) and returns
-                 True to continue or False to stop. May be None.
-
-    Returns:
-        A wrapped callback that raises StopPromptProcessing when stopping
-        is requested. If callback is None, returns None.
-
-    Raises:
-        StopPromptProcessing: When the original callback returns False.
-    """
-    if callback is None:
-        return None
-
-    def inner_callback(percentage: float) -> bool:
-        should_continue = callback(percentage)
-        if not should_continue:
-            logger.info("Prompt processing was cancelled by the user.")
-            raise StopPromptProcessing()
-        return should_continue
-
-    return inner_callback
-
-
-def token_count_callback_wrapper(
-    callback: Optional[Callable[[float], bool]],
-) -> Optional[Callable[[int, int], None]]:
-    """
-    Adapts a float, percentage based progress callback into a token count based one.
-
-    Args:
-        outer_callback: A callback that accepts progress (0.0–100.0) and returns
-                        True to continue or False to stop. May be None.
-
-    Returns:
-        A token-based callback (processed_tokens, total_tokens) -> None,
-        as is expected by mlx-lm's stream_generate.
-        If outer_callback is None, returns None.
-    """
-    if callback is None:
-        return None
-
-    def inner_callback(processed_tokens: int, total_tokens: int) -> None:
-        if total_tokens <= 0:
-            progress = 0.0
-        else:
-            progress = 100 * processed_tokens / total_tokens
-        callback(progress)
-
-    return inner_callback
-
-
 def create_generator(
     model_kit: ModelKit | VisionModelKit,
     prompt_tokens: List[int],
@@ -288,7 +198,7 @@ def create_generator(
     # occurs in different places depending on the model type and prompt content. The prompt will only
     # be processed once, but some contexts are not aware that the prompt is already processed, which
     # can cause the progress to look like it is being reset when it is actually already complete.
-    prompt_progress_callback = ratchet_callback_wrapper(prompt_progress_callback)
+    prompt_progress_callback = ratchet(prompt_progress_callback)
 
     # Set up kv cache
     if type(model_kit) is not VisionModelKit:
@@ -474,9 +384,7 @@ def create_generator(
         prompt=input_tokens,
         max_tokens=max_tokens,
         logits_processors=logits_processors,
-        prompt_progress_callback=token_count_callback_wrapper(
-            throw_to_stop_callback_wrapper(prompt_progress_callback)
-        ),
+        prompt_progress_callback=token_count(throw_to_stop(prompt_progress_callback)),
         prefill_step_size=PROMPT_PROCESSING_CHUNK_SIZE,
         **generate_args,
     )
